+++ src/Symfony/Component/Mailer/Transport.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Mailer;
 
+use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Mailer\Bridge\AhaSend\Transport\AhaSendTransportFactory;
@@ -36,6 +38,7 @@ use Symfony\Component\Mailer\Bridge\Sweego\Transport\SweegoTransportFactory;
 use Symfony\Component\Mailer\Bridge\TurboSmtp\Transport\TurboSmtpTransportFactory;
 use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Exception\UnsupportedSchemeException;
+use Symfony\Component\Mailer\Transport\AbstractTransport;
 use Symfony\Component\Mailer\Transport\Dsn;
 use Symfony\Component\Mailer\Transport\FailoverTransport;
 use Symfony\Component\Mailer\Transport\NativeTransportFactory;
@@ -46,6 +49,7 @@ use Symfony\Component\Mailer\Transport\Smtp\EsmtpTransportFactory;
 use Symfony\Component\Mailer\Transport\TransportFactoryInterface;
 use Symfony\Component\Mailer\Transport\TransportInterface;
 use Symfony\Component\Mailer\Transport\Transports;
+use Symfony\Component\RateLimiter\RateLimiterFactoryInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -97,6 +101,7 @@ final class Transport
      */
     public function __construct(
         private iterable $factories,
+        private ?ContainerInterface $rateLimiterLocator = null,
     ) {
     }
 
@@ -104,17 +109,27 @@ final class Transport
     {
         $transports = [];
         foreach ($dsns as $name => $dsn) {
-            $transports[$name] = $this->fromString($dsn);
+            try {
+                $rateLimiter = $this->rateLimiterLocator?->get($name);
+            } catch (NotFoundExceptionInterface) {
+                $rateLimiter = null;
+            }
+
+            $transports[$name] = $this->fromString($dsn, $rateLimiter);
         }
 
         return new Transports($transports);
     }
 
-    public function fromString(#[\SensitiveParameter] string $dsn): TransportInterface
+    public function fromString(#[\SensitiveParameter] string $dsn, ?RateLimiterFactoryInterface $rateLimiter = null): TransportInterface
     {
         [$transport, $offset] = $this->parseDsn($dsn);
         if ($offset !== \strlen($dsn)) {
             throw new InvalidArgumentException('The mailer DSN has some garbage at the end.');
+        }
+
+        if ($rateLimiter && $transport instanceof AbstractTransport) {
+            $transport->setRateLimiterFactory($rateLimiter);
         }
 
         return $transport;